Remove express from dependencies... hubot already depends on it
Hubot is using v3 of express and we were requiring v4.  Although I'd like for hubot to use v4 so we could be current, the only thing we were using it for was the 'express.static' method which is included in v3 anyway.  Therefore I don't see why we need to require a different version so I've removed it